requirements.txt
PySide6
pyttsx3
SpeechRecognition
pyaudio
python-dotenv
google-generativeai
